Specify page and pageSize on API call

+++ src/Peer.Application/Contracts/PagedList.cs
@@ -3,11 +3,11 @@ public class PagedList<T>
 {
     public long TotalItems { get; private set; }
     public IReadOnlyList<T> Items { get; private set; }
-    public long Page { get; private set; } = 1;
-    public long TotalPages { get; private set; }
+    public int Page { get; private set; } = 1;
+    public int TotalPages { get; private set; }
 
 
-    public PagedList(long totalItems, long page, long totalPages, ICollection<T> items)
+    public PagedList(long totalItems, int page, int totalPages, ICollection<T> items)
     {
         TotalItems = totalItems;
         Items = items.ToList().AsReadOnly();
